# ARROW-2905: [C++] Remove raw_data_ member from BufferBuilder, adaptive int builders