--- a/cpp/src/arrow/buffer-builder.h
+++ b/cpp/src/arrow/buffer-builder.h
@@ -43,7 +43,7 @@ namespace arrow {
 class ARROW_EXPORT BufferBuilder {
  public:
   explicit BufferBuilder(MemoryPool* pool ARROW_MEMORY_POOL_DEFAULT)
-      : pool_(pool), data_(NULLPTR), capacity_(0), size_(0) {}
+      : pool_(pool), capacity_(0), size_(0) {}
 
   /// \brief Resize the buffer to the nearest multiple of 64 bytes
   ///
@@ -64,9 +64,8 @@ class ARROW_EXPORT BufferBuilder {
       ARROW_RETURN_NOT_OK(buffer_->Resize(new_capacity, shrink_to_fit));
     }
     capacity_ = buffer_->capacity();
-    data_ = buffer_->mutable_data();
     if (capacity_ > old_capacity) {
-      memset(data_ + old_capacity, 0, capacity_ - old_capacity);
+      memset(mutable_data() + old_capacity, 0, capacity_ - old_capacity);
     }
     return Status::OK();
   }
@@ -123,7 +122,7 @@ class ARROW_EXPORT BufferBuilder {
   Status Append(const std::array<uint8_t, NBYTES>& data) {
     constexpr auto nbytes = static_cast<int64_t>(NBYTES);
     ARROW_RETURN_NOT_OK(Reserve(NBYTES, true));
-    std::copy(data.cbegin(), data.cend(), data_ + size_);
+    std::copy(data.cbegin(), data.cend(), mutable_data() + size_);
     size_ += nbytes;
     return Status::OK();
   }
@@ -136,12 +135,12 @@ class ARROW_EXPORT BufferBuilder {
 
   // Unsafe methods don't check existing size
   void UnsafeAppend(const void* data, const int64_t length) {
-    memcpy(data_ + size_, data, static_cast<size_t>(length));
+    memcpy(mutable_data() + size_, data, static_cast<size_t>(length));
     size_ += length;
   }
 
   void UnsafeAppend(const int64_t num_copies, uint8_t value) {
-    memset(data_ + size_, value, static_cast<size_t>(num_copies));
+    memset(mutable_data() + size_, value, static_cast<size_t>(num_copies));
     size_ += num_copies;
   }
 
@@ -169,13 +168,15 @@ class ARROW_EXPORT BufferBuilder {
 
   int64_t capacity() const { return capacity_; }
   int64_t length() const { return size_; }
-  const uint8_t* data() const { return data_; }
-  uint8_t* mutable_data() { return data_; }
+  const uint8_t* data() const { return buffer_->data(); }
+  uint8_t* mutable_data() {
+    return buffer_->mutable_data();
+    ;
+  }
 
  private:
   std::shared_ptr<ResizableBuffer> buffer_;
   MemoryPool* pool_;
-  uint8_t* data_;
   int64_t capacity_;
   int64_t size_;
 };