Removing this member had no appreciable effect on performance in the builder benchmark:

```
Before
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                     7910 us       7910 us         89   7.90159GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_stddev                     93 us         93 us         89   95.2444MB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean        16794 us      16794 us         42   3.72172GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_stddev        144 us        144 us         42   32.5893MB/s

After
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                     7782 us       7781 us         89   8.03249GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_stddev                     54 us         53 us         89    55.593MB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean        16589 us      16588 us         42   3.76803GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_stddev        195 us        194 us         42   45.0538MB/s
```