@wesm 2019-03-11
Since ARROW-2905 was opened a good bit of logic is now centralized in buffer-builder.h. There are still raw pointers in use there. Do they impact builder performance? As I recall the prior discussion suggested that the behavior may have changed from gcc 4 to gcc 5+