@wesm 2019-03-12
I can't figure out how to disable Turbo Boost on my Thinkpad P1 so take the benchmarks with a grain of salt, but I don't see an appreciable difference so that's good =) Some of the noise seems to be due to memory allocation

before:

```
------------------------------------------------------------------------------------------------------
Benchmark                                                               Time           CPU Iterations
------------------------------------------------------------------------------------------------------
BM_BuildPrimitiveArrayNoNulls/min_time:1.000                        97965 us      97979 us         14   5.10316GB/s
BM_BuildPrimitiveArrayNoNullsScalarUnsafeAppend/min_time:5.000     213374 us     213379 us         32   2.34325GB/s
BM_BuildPrimitiveArrayNoNullsScalarSafeAppend/min_time:5.000       205611 us     205594 us         33   2.43198GB/s
```

after:

```
------------------------------------------------------------------------------------------------------
Benchmark                                                               Time           CPU Iterations
------------------------------------------------------------------------------------------------------
BM_BuildPrimitiveArrayNoNulls/min_time:1.000                        99425 us      99389 us         14   5.03073GB/s
BM_BuildPrimitiveArrayNoNullsScalarUnsafeAppend/min_time:5.000     206839 us     206790 us         34   2.41791GB/s
BM_BuildPrimitiveArrayNoNullsScalarSafeAppend/min_time:5.000       205519 us     205496 us         34   2.43313GB/s
```

BTW I have started normalizing microbenchmarks to use `MinTime(1.0)` or similar. I think this may yield better quality / more consistent results than "repeats"