@wesm 2019-03-17
It's hard to really tell whether this is causing an impact or not. The most affected benchmarks by this change are the BinaryBuilder benchmarks, because they append 10-length strings to create an approx 160MB data structure. I increased these benchmarks to run with MinTime 10 seconds, made sure to disable CPU throttling, and I'm showing roughly 1-3% perf degradation

before:

```
------------------------------------------------------------------------------------
Benchmark                                             Time           CPU Iterations
------------------------------------------------------------------------------------
BM_BuildBinaryArray/min_time:10.000              495018 us     494879 us         28   323.311MB/s
BM_BuildChunkedBinaryArray/min_time:10.000       370306 us     370071 us         37    432.35MB/s
BM_BuildFixedSizeBinaryArray/min_time:10.000      10139 us      10138 us       1000   986.411MB/s
```

after

```
------------------------------------------------------------------------------------
Benchmark                                             Time           CPU Iterations
------------------------------------------------------------------------------------
BM_BuildBinaryArray/min_time:10.000              496589 us     496595 us         28   322.194MB/s
BM_BuildChunkedBinaryArray/min_time:10.000       375714 us     375717 us         37   425.852MB/s
BM_BuildFixedSizeBinaryArray/min_time:10.000      10489 us      10489 us       1000   953.402MB/s
```

This is probably much ado about nothing, but I'm concerned that we lack appropriate benchmarking infrastructure to evaluate this decision. Since there's an effort to establish better benchmarking systems in the next couple of months, I recommend we do not merge this patch now and wait until we can get more data points (using this patch to help determine whether we have a rigorous process in place to determine whether a patch creates a statistically significant performance improvement or regression)